@@ -19,7 +19,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 2.0.1+1
 
 environment:
-  sdk: ^3.6.0
+  sdk: ^3.8.0
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
 # consider running `flutter pub upgrade --major-versions`. Alternatively,
@@ -131,7 +131,7 @@ dev_dependencies:
   flutter_lints: ^5.0.0
   build_runner: ^2.0.0
   freezed: ^3.0.1
-  json_serializable: any
+  json_serializable: ^6.7.1
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
